🔥 Remove: VerifyChallege 코드를 VerifyPage로 이동
