+++ src/pages/verifyPage/VerifyPage.module.scss
@@ -1,5 +1,5 @@
-@use '../../../../styles/mixins' as m;
-@use '../../../../styles/variables' as *;
+@use '../../styles/mixins' as m;
+@use '../../styles/variables' as *;
 
 .verifyChallenge {
   @include m.flex-column-center(flex-start, center, 8px);
